moved EnablePprof from backend.go to setEthconfig in flags.go + enabled block + mutex
